# docs/api-specs/openapi.yaml
openapi: 3.0.3
info:
  title: Second Brain AI Companion API
  description: API for the Second Brain AI Companion system
  version: 1.0.0
  contact:
    name: API Support
    email: support@secondbrain.ai

servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.secondbrain.ai/v1
    description: Production server

paths:
  /documents/upload:
    post:
      summary: Upload and process a document
      description: Upload a file for processing and ingestion into the knowledge base
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
                tags:
                  type: array
                  items:
                    type: string
                  description: Optional tags to associate with the document
                metadata:
                  type: object
                  description: Additional metadata for the document
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          description: Invalid file or request
        '413':
          description: File too large
        '422':
          description: Unsupported file type

  /documents:
    get:
      summary: List user documents
      description: Retrieve a paginated list of user's documents
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: source_type
          in: query
          schema:
            type: string
            enum: [audio, pdf, web, text, image]
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'

  /documents/{document_id}:
    get:
      summary: Get document details
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          description: Document not found

    delete:
      summary: Delete a document
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document deleted successfully
        '404':
          description: Document not found

  /query:
    post:
      summary: Submit a natural language query
      description: Submit a query to the AI companion and get a response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Query response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Invalid query

  /conversations:
    get:
      summary: Get conversation history
      description: Retrieve user's conversation history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Conversation history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'

  /search:
    get:
      summary: Advanced search
      description: Perform advanced search with filters
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
        - name: source_type
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [audio, pdf, web, text, image]
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'

  /tags:
    get:
      summary: List all tags
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'

    post:
      summary: Create a new tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'

  /analytics:
    get:
      summary: Get usage analytics
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'

components:
  schemas:
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source_path:
          type: string
        source_type:
          type: string
          enum: [audio, pdf, web, text, image]
        title:
          type: string
        author:
          type: string
        created_at:
          type: string
          format: date-time
        ingested_at:
          type: string
          format: date-time
        file_size:
          type: integer
        processing_status:
          type: string
          enum: [pending, processing, completed, failed]
        metadata:
          type: object
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        chunk_count:
          type: integer

    DocumentListResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentResponse'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        has_next:
          type: boolean

    QueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The natural language query
        conversation_id:
          type: string
          format: uuid
          description: Optional conversation ID for context
        filters:
          type: object
          properties:
            source_types:
              type: array
              items:
                type: string
            date_range:
              type: object
              properties:
                start:
                  type: string
                  format: date
                end:
                  type: string
                  format: date
            tags:
              type: array
              items:
                type: string

    QueryResponse:
      type: object
      properties:
        conversation_id:
          type: string
          format: uuid
        response:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SourceReference'
        response_time_ms:
          type: integer
        created_at:
          type: string
          format: date-time

    SourceReference:
      type: object
      properties:
        document_id:
          type: string
          format: uuid
        chunk_id:
          type: string
          format: uuid
        title:
          type: string
        excerpt:
          type: string
        relevance_score:
          type: number
          format: float
        source_type:
          type: string

    ConversationListResponse:
      type: object
      properties:
        conversations:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              query:
                type: string
              response:
                type: string
              created_at:
                type: string
                format: date-time
              response_time_ms:
                type: integer
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              document_id:
                type: string
                format: uuid
              chunk_id:
                type: string
                format: uuid
              title:
                type: string
              content:
                type: string
              relevance_score:
                type: number
                format: float
              source_type:
                type: string
              created_at:
                type: string
                format: date-time
        total:
          type: integer
        query_time_ms:
          type: integer

    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        auto_generated:
          type: boolean

    CreateTagRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          default: '#3B82F6'

    AnalyticsResponse:
      type: object
      properties:
        total_documents:
          type: integer
        total_chunks:
          type: integer
        total_queries:
          type: integer
        avg_response_time_ms:
          type: number
          format: float
        storage_usage_mb:
          type: number
          format: float
        popular_tags:
          type: array
          items:
            type: object
            properties:
              tag:
                type: string
              count:
                type: integer
        query_trends:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              query_count:
                type: integer

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - BearerAuth: []